+++ src/pages/vendorDashboard/vendorProducts.tsx
@@ -78,7 +78,13 @@ const vendorProducts = () => {
     fetchData();
   }, []);
 
-   
+   if (isLoading) {
+        return <div className="w-8 h-8 border-b-2 border-gray-900 rounded-full animate-spin mx-auto my-auto"></div>;
+    }
+
+    if (error) {
+        return <div>Error: {error}</div>;
+    }
 
   return (
     <div className="flex-1 overflow-auto relative z-10">
